requirements.txt
click==6.6
docutils==0.12
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.3.0
shortid==0.1.0
SQLAlchemy==1.0.14
Werkzeug==0.11.10
